refactor: centralize type definitions in types folder
- Move Service interface from data/services.ts to types/what-we-do.ts
- Move CaseStudy interface from data/case-studies.ts to types/what-we-do.ts
- Update all import statements across components, utils, and tests
- Ensure consistent pattern: data files contain data, types files contain types
- Fix TypeScript compilation errors and improve maintainability

This creates a cleaner architecture with centralized type management and
better separation of concerns between data and type definitions.
