+++ components/case-studies/case-study-card.tsx
@@ -3,7 +3,7 @@
 import React from 'react'
 import Image from 'next/image'
 import Link from 'next/link'
-import type { CaseStudy } from '@/data/case-studies'
+import type { CaseStudy } from '@/types/what-we-do'
 
 export interface CaseStudyCardProps {
   caseStudy: CaseStudy
