+++ types/what-we-do.ts
@@ -1,5 +1,36 @@
-import type { CaseStudy } from '@/data/case-studies'
-import type { Service } from '@/data/services'
+export interface CaseStudy {
+  id: string
+  title: string
+  serviceIds: string[]
+  shortDescription: string
+  detailedDescription: string
+  results: string[]
+  imageUrl: string
+  client: string
+  industryIds: string[]
+  duration: string
+  technologies: string[]
+}
+
+export interface Service {
+  id: string
+  title: string
+  shortDescription: string
+  detailedDescription: string
+  icon: string // Path to the icon/image
+  benefits: string[]
+  process?: {
+    steps: {
+      title: string
+      description: string
+    }[]
+  }
+  deliverables?: string[]
+  caseStudies?: {
+    id: string
+    title: string
+  }[] // For future integration with case studies
+}
 
 export interface Industry {
   id: string
